go: only track the latest version
There is no much value in making old versions available as they have not been
updated regularly.

In reality, it is not even a good practice not to use the latest version available,
so the best thing to do would be drop the old versions and only keep traking of the
latest version in oe-core master branch.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>