oeqa/selftest/rust: disable codegen/i128-x86-align.rs test
Re-disable the codegen/i128-x86-align.rs test for now, as it is
failing when testing on qemux86-64. Further investigation into the
testcase is required, as it is believed the relevant changes for it
to work have already been backported (i.e. matching rust-llvm and
the layout fix from commit 96f11d29).

Signed-off-by: Scott Murray <scott.murray@konsulko.com>